Don't define named classes in anonymous class
